Move `WindowContextEvent` to `backend` mod
