+++ crates/wolf_engine_window/src/event.rs
@@ -4,7 +4,7 @@ use uuid::Uuid;
 use wolf_engine_events::dynamic::Event;
 use wolf_engine_input::Input;
 
-use crate::{error::WindowError, WindowSettings};
+use crate::error::WindowError;
 
 /// An event generated by the window system.
 #[derive(Clone, PartialEq, Debug)]
@@ -21,15 +21,3 @@ pub enum WindowEvent {
 }
 
 impl Event for WindowEvent {}
-
-#[derive(Clone, PartialEq, Debug)]
-pub enum WindowContextEvent {
-    WindowRenameRequested(Uuid, String),
-    WindowRedrawRequested(Uuid),
-    WindowResized(Uuid, u32, u32),
-    WindowCreated(Uuid, WindowSettings),
-    WindowClosed(Uuid),
-    Exited,
-}
-
-impl Event for WindowContextEvent {}
